requirements.txt
fastapi
pydantic
uvicorn[standard]
passlib
python-jose[cryptography]
bcrypt<4.1
python-multipart
requests
python-dotenv
currentsapi
serpapi
langchain
langchain_core
langchain_openai
pydantic_settings
rich
google-search-results
sqlalchemy